+++ CHANGES.rst
@@ -3,7 +3,7 @@ Products.MeetingLalouviere Changelog
 
 The Products.MeetingCommunes version must be the same as the Products.PloneMeeting version
 
-4.2.15 (unreleased)
+4.2.15 (2026-05-21)
 -------------------
 
 - Do not keep fields `interventions`, `committeeTranscript` when cloning item.